requirements.txt
flask>=3.0
httpx>=0.28
huggingface_hub>=1.0
Pillow>=10.0
instaloader>=4.10
instagrapi>=2.1,<2.2
duckduckgo-search>=6.0
gunicorn>=22.0
python-dotenv>=1.0
